Widen Diffuser memory-budget stepCount to uint64_t (#31)
Replace a narrowing std::uint32_t cast of diffuser.steps.size() at the
validateResolvedConfig call site with a signature-level fix: widen
estimateDiffuserMemoryBytes/checkDiffuserMemoryBudget's stepCount
parameter to std::uint64_t, matching what std::vector::size() returns
and what the checked-arithmetic helpers already operate in. This
removes the truncation risk at the type level instead of adding a
bespoke bounds check at one call site; the existing overflow-checked
multiplication already rejects any resulting absurd byte estimate.

+++ src/config/ResolveConfig.cpp
@@ -117,7 +117,7 @@ std::optional<std::uint64_t> checkedAdd(
 std::optional<std::uint64_t> estimateDiffuserMemoryBytes(
     const std::uint32_t channels,
     const std::uint64_t totalSamples,
-    const std::uint32_t stepCount) noexcept {
+    const std::uint64_t stepCount) noexcept {
   constexpr std::uint64_t kSampleBytes = 8;
   constexpr std::uint64_t kMetadataBytesPerChannel = 24;
   auto delayBytes = checkedMul(channels, totalSamples);
@@ -147,7 +147,7 @@ std::optional<std::uint64_t> estimateDiffuserMemoryBytes(
 void checkDiffuserMemoryBudget(
     const std::uint32_t channels,
     const std::uint64_t totalSamples,
-    const std::uint32_t stepCount,
+    const std::uint64_t stepCount,
     const std::uint64_t budgetBytes,
     const std::string_view path) {
   const auto estimate =
@@ -778,7 +778,7 @@ void validateResolvedConfig(
   checkDiffuserMemoryBudget(
       channels,
       diffuser.totalSamples,
-      static_cast<std::uint32_t>(diffuser.steps.size()),
+      diffuser.steps.size(),
       memoryBudgetBytes,
       "/composition/stages/1");
   if (validatingRequest &&
